Try to fix PR1 travis CI
There is a complex dependency loop for windows installations

Archive -> File[$install_dir] -> File["C:\Program Files"] (implicit!)

At the same time, the archive{} has the attributes "creates" on
$version_file. $version file lifes within $install_dir.

Thus we do not need to those explicit dependencies, I think.

I'm not sure wether it is enough to rely on "pdk test unit" here and I
do not have any windows systems to test this on, so please test this
somewhere and give feedback wether the archive -> creates assumption is
correct.